pubsub: connect to peers before a 1-way pub
pubsub: tmp increase timeout

pubsub: tmp block before publish

pubsub: filter peer info by topic
