+++ ipfs/swarm.go
@@ -1,0 +1,138 @@
+package ipfs
+
+import (
+	"context"
+	"fmt"
+	"sort"
+
+	"gx/ipfs/QmPDEJTb3WBHmvubsLXCaqRPC8dRgvFz7A4p96dxZbJuWL/go-ipfs/core"
+	"gx/ipfs/QmPDEJTb3WBHmvubsLXCaqRPC8dRgvFz7A4p96dxZbJuWL/go-ipfs/core/coreapi"
+	inet "gx/ipfs/QmY3ArotKMKaL7YGfbQfyDrib6RVraLqZYWXZvVgZktBxp/go-libp2p-net"
+)
+
+// SwarmConnect opens a direct connection to a list of peer multi addresses
+func SwarmConnect(node *core.IpfsNode, addrs []string) ([]string, error) {
+	api, err := coreapi.NewCoreAPI(node)
+	if err != nil {
+		return nil, err
+	}
+
+	pis, err := peersWithAddresses(addrs)
+	if err != nil {
+		return nil, err
+	}
+
+	ctx, cancel := context.WithTimeout(node.Context(), connectTimeout)
+	defer cancel()
+
+	output := make([]string, len(pis))
+	for i, pi := range pis {
+		output[i] = "connect " + pi.ID.Pretty()
+
+		err := api.Swarm().Connect(ctx, pi)
+		if err != nil {
+			return nil, fmt.Errorf("%s failure: %s", output[i], err)
+		}
+		output[i] += " success"
+	}
+
+	return output, nil
+}
+
+type streamInfo struct {
+	Protocol string
+}
+
+type connInfo struct {
+	Addr      string         `json:"addr"`
+	Peer      string         `json:"peer"`
+	Latency   string         `json:"latency,omitempty"`
+	Muxer     string         `json:"muxer,omitempty"`
+	Direction inet.Direction `json:"direction,omitempty"`
+	Streams   []streamInfo   `json:"streams,omitempty"`
+}
+
+func (ci *connInfo) Less(i, j int) bool {
+	return ci.Streams[i].Protocol < ci.Streams[j].Protocol
+}
+
+func (ci *connInfo) Len() int {
+	return len(ci.Streams)
+}
+
+func (ci *connInfo) Swap(i, j int) {
+	ci.Streams[i], ci.Streams[j] = ci.Streams[j], ci.Streams[i]
+}
+
+type ConnInfos struct {
+	Peers []connInfo
+}
+
+func (ci ConnInfos) Less(i, j int) bool {
+	return ci.Peers[i].Addr < ci.Peers[j].Addr
+}
+
+func (ci ConnInfos) Len() int {
+	return len(ci.Peers)
+}
+
+func (ci ConnInfos) Swap(i, j int) {
+	ci.Peers[i], ci.Peers[j] = ci.Peers[j], ci.Peers[i]
+}
+
+// SwarmPeers lists the set of peers this node is connected to
+func SwarmPeers(node *core.IpfsNode, verbose bool, latency bool, streams bool, direction bool) (*ConnInfos, error) {
+	api, err := coreapi.NewCoreAPI(node)
+	if err != nil {
+		return nil, err
+	}
+
+	ctx, cancel := context.WithTimeout(node.Context(), connectTimeout)
+	defer cancel()
+
+	conns, err := api.Swarm().Peers(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	var out ConnInfos
+	for _, c := range conns {
+		ci := connInfo{
+			Addr: c.Address().String(),
+			Peer: c.ID().Pretty(),
+		}
+
+		if verbose || direction {
+			// set direction
+			ci.Direction = c.Direction()
+		}
+
+		if verbose || latency {
+			lat, err := c.Latency()
+			if err != nil {
+				return nil, err
+			}
+
+			if lat == 0 {
+				ci.Latency = "n/a"
+			} else {
+				ci.Latency = lat.String()
+			}
+		}
+		if verbose || streams {
+			strs, err := c.Streams()
+			if err != nil {
+				return nil, err
+			}
+
+			for _, s := range strs {
+				ci.Streams = append(ci.Streams, streamInfo{Protocol: string(s)})
+			}
+		}
+		sort.Sort(&ci)
+		out.Peers = append(out.Peers, ci)
+	}
+
+	sort.Sort(&out)
+	return &out, nil
+}
