+++ ipfs/util.go
@@ -11,9 +11,16 @@ import (
 	ma "gx/ipfs/QmTZBfrPJmjWsCvHEtX5FE6KimVJhsJg5sBbqEFYf4UZtL/go-multiaddr"
 	"gx/ipfs/QmYVXrKrKHDC9FobgmcmshCDyWwdrfwfanNQN4oxJ9Fk3h/go-libp2p-peer"
 	pstore "gx/ipfs/QmaCTz9RkrU13bm9kMB54f7atgqM4qkjDZpRwRoJiWXEqs/go-libp2p-peerstore"
-	"gx/ipfs/QmbeHtaBy9nZsW4cHRcvgVY4CnDhXudE2Dr6qDxS7yg9rX/go-libp2p-record"
 	"gx/ipfs/Qmdf1djucJ1jX5RMF1bDbFg5ybZnupmSAeETQQ3ZV7z6dU/go-ipfs-addr"
 )
+
+// ShortenID returns the last 7 chars of a string
+func ShortenID(id string) string {
+	if len(id) < 7 {
+		return id
+	}
+	return id[len(id)-7:]
+}
 
 // PrintSwarmAddrs prints the addresses of the host
 func PrintSwarmAddrs(node *core.IpfsNode) error {
@@ -68,36 +75,6 @@ func PublicIPv4Addr(node *core.IpfsNode) (string, error) {
 		return pub, errors.New("no public ipv4 address found")
 	}
 	return pub, nil
-}
-
-// ShortenID returns the last 7 chars of a string
-func ShortenID(id string) string {
-	if len(id) < 7 {
-		return id
-	}
-	return id[len(id)-7:]
-}
-
-// IpnsSubs shows current name subscriptions
-func IpnsSubs(node *core.IpfsNode) ([]string, error) {
-	if node.PSRouter == nil {
-		return nil, errors.New("IPNS pubsub subsystem is not enabled")
-	}
-	var paths []string
-	for _, key := range node.PSRouter.GetSubscriptions() {
-		ns, k, err := record.SplitKey(key)
-		if err != nil || ns != "ipns" {
-			// not necessarily an error.
-			continue
-		}
-		pid, err := peer.IDFromString(k)
-		if err != nil {
-			log.Errorf("ipns key not a valid peer ID: %s", err)
-			continue
-		}
-		paths = append(paths, "/ipns/"+peer.IDB58Encode(pid))
-	}
-	return paths, nil
 }
 
 // parseAddresses is a function that takes in a slice of string peer addresses
