+++ service/main.go
@@ -594,56 +594,3 @@ func (srv *Service) listen(tag string) {
 		}
 	}
 }
-
-// ListenFor opens a tmp subscription to a topic and passes results to handler
-//func (srv *Service) ListenFor(topic string, discover bool, handler func(pid peer.ID, env *pb.Envelope) (*pb.Envelope, error)) context.CancelFunc {
-//	msgs := make(chan iface.PubSubMessage, 10)
-//	ctx, cancel := context.WithCancel(srv.Node().Context())
-//	go func() {
-//		if err := ipfs.Subscribe(srv.Node(), ctx, topic, discover, msgs); err != nil {
-//			close(msgs)
-//			log.Errorf("pubsub listener stopped with error: %s", err)
-//			return
-//		}
-//	}()
-//	log.Debugf("pubsub listener started for %s", topic)
-//
-//	go func() {
-//		for {
-//			select {
-//			// end loop on context close
-//			case <-ctx.Done():
-//				log.Debugf("pubsub listener shutdown for %s", topic)
-//				return
-//			case msg, ok := <-msgs:
-//				if !ok {
-//					log.Debugf("pubsub listener shutdown for %s", topic)
-//					return
-//				}
-//
-//				mPeer := msg.From()
-//				if mPeer.Pretty() == srv.Node().Identity.Pretty() {
-//					continue
-//				}
-//
-//				pmes := new(pb.Envelope)
-//				if err := proto.Unmarshal(msg.Data(), pmes); err != nil {
-//					log.Errorf("error unmarshaling pubsub message data from %s: %s", mPeer.Pretty(), err)
-//					continue
-//				}
-//
-//				if err := srv.VerifyEnvelope(pmes, mPeer); err != nil {
-//					log.Warningf("error verifying message: %s", err)
-//					continue
-//				}
-//
-//				log.Debugf("received pubsub %s from %s", pmes.Message.Type.String(), mPeer.Pretty())
-//				if _, err := handler(mPeer, pmes); err != nil {
-//					log.Errorf("%s handle message error: %s", pmes.Message.Type.String(), err)
-//				}
-//			}
-//		}
-//	}()
-//
-//	return cancel
-//}
